refactor: move Draggable to entities layer
