+++ src/entities/drag-n-drop/Draggable.ts
@@ -1,4 +1,4 @@
-import Component from "./ui/base/Component";
+import Component from "../../shared/ui/base/Component";
 
 const DRAG_THRESHOLD = 10;
