+++ src/entities/drag-n-drop/index.ts
@@ -1,0 +1,3 @@
+import Draggable from "./Draggable";
+
+export default Draggable;
